Move adapters to their own classes
