+++ lib/nochmal/reupload.rb
@@ -9,7 +9,7 @@ module Nochmal
     attr_reader :active_storage, :from_service, :to_service
 
     def initialize(from:, to: nil, helper: nil)
-      @active_storage = helper || ActiveStorageHelper.new
+      @active_storage = helper || Adapters::ActiveStorage.new
       @from_service = active_storage.from_storage_service(from.to_sym)
       @to_service = active_storage.to_storage_service(to&.to_sym)
       @notes = []
